refactor: split dashboard metric card into flip card components
